02-twitter | setting, Firebase Auth 사용을 위한 설정 #41
- v8.0이상 버전으로 최신 문법 적용
