+++ 02.twitter-clone/nwitter/src/components/App.js
@@ -1,8 +1,15 @@
-import React from 'react';
+import React, { useState } from 'react';
 import AppRouter from './Router';
+import { authService } from '../fbase';
 
 function App() {
-  return <AppRouter />;
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  return (
+    <>
+      <AppRouter isLoggedIn={isLoggedIn} />
+      <footer>&copy; {new Date().getFullYear()} Nwitter</footer>
+    </>
+  );
 }
 
 export default App;
