+++ 02.twitter-clone/nwitter/src/components/Router.js
@@ -3,8 +3,7 @@ import { HashRouter as Router, Route, Routes } from 'react-router-dom';
 import Auth from '../routes/Auth';
 import Home from '../routes/Home';
 
-const AppRouter = () => {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
+const AppRouter = ({ isLoggedIn }) => {
   return (
     <Router>
       <Routes>
