+++ 02.twitter-clone/nwitter/src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/App';
-import firebase from './firebase';
+import { authService } from './fbase';
 
-console.log(firebase);
+console.log(authService);
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
